Boucle shell amelioree.

+++ tsh.c
@@ -11,31 +11,27 @@ int main(void){
     char commande[56];
     char *newArgv[] = {NULL};
     char *envp[] = {NULL};
-    
     while(1){
         printf("tsh> ");
-        
         char *exe = fgets(commande, sizeof(commande), stdin);
         
         if(exe != NULL){
             exe[strlen(commande) - 1] = '\0';
-            id = fork();
-
             if(!strcmp(commande, "exit")) exit(0);
-
-            if(id != 0){
-                printf("Parent: %d \n", getpid());
-                wait(NULL);
-            }else{
-                printf("Enfant: %d\n", getpid());
-                if(execve(commande, newArgv, envp) == -1){
-                    kill(getpid(), SIGKILL);
-                }
-            }
-            //free(commande);
-            //system("ps");
-            //system("ps");
             
+            if(strcmp(exe, "\0")){
+                id = fork();
+
+                if(id != 0){
+                    wait(NULL);
+
+                }else{
+                    if(execve(commande, newArgv, envp) == -1){
+                        printf("Commande introuvable.\n");
+                        kill(getpid(), SIGKILL);
+                    }
+                } 
+            }
         }
     }
     printf("Process %d terminating.\n", getpid());
